[native] Fix GROUPING SETS with multiple aliased columns
GROUPING SETS can be specified with multiple alias columns of the
same input column. This is typically used to compute multiple
mixed aggregations on the same key.

Presto built plan fragments correctly specifying the output column
name for such cases. But the Velox plan conversion always mapped
to the input column losing the distinction of the output group keys.